MNT: Move customized dict classes to utils.py.
MNT: Update imports.
